fix(mgn): self-enforce int64 edge indices; rename tf alias to F
cells_to_edges now coerces its torch.unique output to int64 so an int32
cells input can't silently produce an int32 edge index (which would
raise downstream in index_add_), mirroring world_edges' existing
guarantee. Also rename the torch.nn.functional import alias in
simulator.py from tf to F, since tf reads as TensorFlow in a module
whose plan pins "No TensorFlow"; no behavior change.

+++ src/structbench/models/mgn/mesh_ops.py
@@ -31,7 +31,7 @@ def cells_to_edges(cells: Tensor) -> Tensor:
     dst = torch.cat([cells[:, b] for a, b in pairs])
     und = torch.stack([torch.cat([src, dst]), torch.cat([dst, src])])  # (2, 2*n)
     und = und[:, und[0] != und[1]]
-    return torch.unique(und, dim=1)
+    return torch.unique(und, dim=1).to(torch.int64)
 
 
 def world_edges(positions: Tensor, radius: float, mesh_edge_index: Tensor) -> Tensor:
